--- android/app/src/main/java/com/example/fire_detector/ui/alarm/AlarmItem.kt
@@ -1,8 +1,0 @@
-package com.example.fire_detector.ui.alarm
-
-data class AlarmItem(
-    val id: Long,
-    val title: String,
-    val message: String,
-    val time: String
-)
