--- android/app/src/main/java/com/example/fire_detector/ui/alarm/AlarmScreen.kt
@@ -1,84 +1,0 @@
-package com.example.fire_detector.ui.alarm
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Close
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AlarmScreen(
-    alarms: List<AlarmItem>,
-    onClose: () -> Unit,
-    onAlarmClick: () -> Unit   // 알림 클릭 → 대시보드
-) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                modifier = Modifier.background(
-                    Brush.horizontalGradient(
-                        listOf(
-                            MaterialTheme.colorScheme.primary,
-                            MaterialTheme.colorScheme.secondary
-                        )
-                    )
-                ),
-                title = {
-                    Text(
-                        text = "알림",
-                        color = Color.White
-                    )
-                },
-                actions = {
-                    IconButton(onClick = onClose) {
-                        Icon(
-                            imageVector = Icons.Outlined.Close,
-                            contentDescription = "닫기",
-                            tint = Color.White
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = Color.Transparent
-                )
-            )
-        }
-    ) { paddingValues ->
-
-        if (alarms.isEmpty()) {
-            EmptyAlarmView(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(paddingValues)
-            )
-        } else {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(paddingValues)
-            ) {
-                items(alarms) { alarm ->
-                    AlarmItemView(
-                        alarm = alarm,
-                        onClick = onAlarmClick
-                    )
-                }
-            }
-        }
-    }
-}
