--- android/app/src/main/java/com/example/fire_detector/ui/splash/SplashScreen.kt
@@ -1,79 +1,0 @@
-package com.example.fire_detector.ui.splash
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import kotlinx.coroutines.delay
-import com.example.fire_detector.ui.icons.Fire
-
-@Composable
-fun SplashScreen(
-    onResult: (hasUrl: Boolean) -> Unit,
-    durationMillis: Long = 1500L
-) {
-    LaunchedEffect(Unit) {
-        delay(durationMillis)
-        onResult(true) // 실제 판단은 NavHost에서
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(
-                Brush.linearGradient(
-                    listOf(
-                        MaterialTheme.colorScheme.primary,
-                        MaterialTheme.colorScheme.secondary
-                    )
-                )
-            ),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-
-            Box(
-                modifier = Modifier
-                    .size(96.dp)
-                    .clip(CircleShape)
-                    .background(Color.White.copy(alpha = 0.22f)),
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    imageVector = Fire,
-                    contentDescription = "Fire",
-                    tint = Color.White,
-                    modifier = Modifier.size(54.dp)
-                )
-            }
-
-            Spacer(modifier = Modifier.height(22.dp))
-
-            Text(
-                text = "Fire Detector",
-                color = Color.White,
-                fontSize = 28.sp,
-                fontWeight = FontWeight.SemiBold
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            Text(
-                text = "화재 감지 시스템",
-                color = Color.White.copy(alpha = 0.9f),
-                fontSize = 14.sp
-            )
-        }
-    }
-}
