+++ android/app/src/main/java/com/hallymUniv/fire_detector/fcm/FireFirebaseMessagingService.kt
@@ -1,0 +1,23 @@
+package com.hallymUniv.fire_detector.fcm
+
+import android.util.Log
+import com.google.firebase.messaging.FirebaseMessagingService
+import com.google.firebase.messaging.RemoteMessage
+
+class FireFirebaseMessagingService : FirebaseMessagingService() {
+
+    override fun onNewToken(token: String) {
+        super.onNewToken(token)
+        Log.d("FCM", "New token: $token")
+
+        // TODO: 서버에 token 전달 (나중에)
+    }
+
+    override fun onMessageReceived(message: RemoteMessage) {
+        super.onMessageReceived(message)
+
+        Log.d("FCM", "From: ${message.from}")
+        Log.d("FCM", "Data: ${message.data}")
+        Log.d("FCM", "Notification: ${message.notification?.title}")
+    }
+}
