+++ android/app/src/main/java/com/hallymUniv/fire_detector/ui/alarm/AlarmItem.kt
@@ -1,0 +1,8 @@
+package com.hallymUniv.fire_detector.ui.alarm
+
+data class AlarmItem(
+    val id: Long,
+    val title: String,
+    val message: String,
+    val time: String
+)
