+++ android/app/src/main/java/com/hallymUniv/fire_detector/ui/alarm/EmptyAlarmView.kt
@@ -1,0 +1,26 @@
+package com.hallymUniv.fire_detector.ui.alarm
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun EmptyAlarmView(
+    modifier: Modifier = Modifier
+) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = "알림이 없습니다",
+            color = Color.Gray,
+            style = MaterialTheme.typography.bodySmall
+        )
+    }
+}
