+++ android/app/src/main/java/com/hallymUniv/fire_detector/ui/dashboard/DashboardScreen.kt
@@ -1,0 +1,79 @@
+package com.hallymUniv.fire_detector.ui.dashboard
+
+import android.webkit.WebView
+import android.webkit.WebViewClient
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Notifications
+import androidx.compose.material.icons.outlined.Settings
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.viewinterop.AndroidView
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DashboardScreen(
+    dashboardUrl: String,
+    onAlarmClick: () -> Unit,
+    onSettingClick: () -> Unit
+) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                modifier = Modifier
+                    .background(
+                        Brush.horizontalGradient(
+                            colors = listOf(
+                                MaterialTheme.colorScheme.primary,
+                                MaterialTheme.colorScheme.secondary
+                            )
+                        )
+                    ),
+                title = {
+                    Text(
+                        text = "Fire Detector",
+                        color = Color.White
+                    )
+                },
+                actions = {
+                    IconButton(onClick = onAlarmClick) {
+                        Icon(
+                            imageVector = Icons.Outlined.Notifications,
+                            contentDescription = "알림",
+                            tint = Color.White
+                        )
+                    }
+                    IconButton(onClick = onSettingClick) {
+                        Icon(
+                            imageVector = Icons.Outlined.Settings,
+                            contentDescription = "설정",
+                            tint = Color.White
+                        )
+                    }
+                },
+                colors = TopAppBarDefaults.topAppBarColors(
+                    containerColor = Color.Transparent
+                )
+            )
+        }
+    ) { paddingValues ->
+        AndroidView(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues),
+            factory = { context ->
+                WebView(context).apply {
+                    settings.javaScriptEnabled = true
+                    settings.domStorageEnabled = true
+                    webViewClient = WebViewClient()
+                    loadUrl(dashboardUrl)
+                }
+            }
+        )
+    }
+}
