+++ android/app/src/main/java/com/hallymUniv/fire_detector/ui/theme/Color.kt
@@ -1,0 +1,6 @@
+package com.hallymUniv.fire_detector.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val FireOrange = Color(0xFFFF6A00)
+val FireRed = Color(0xFFFF0000)
